refactor(plugins): add requestRoomSelection util
