+++ src/maps.js
@@ -4,14 +4,14 @@
  */
 
 import escapeHtml from 'escape-html'
-import Vue from 'vue'
 
 import { getRequestToken } from '@nextcloud/auth'
 import { showSuccess, showError } from '@nextcloud/dialogs'
-import { t, n } from '@nextcloud/l10n'
+import { t } from '@nextcloud/l10n'
 import { generateFilePath, generateUrl } from '@nextcloud/router'
 
 import { postRichObjectToConversation } from './services/messagesService.ts'
+import { requestRoomSelection } from './utils/requestRoomSelection.js'
 
 import '@nextcloud/dialogs/style.css'
 
@@ -33,9 +33,9 @@ async function postLocationToRoom(location, { token, displayName }) {
 
 		showSuccess(t('spreed', 'Location has been posted to {conversation}')
 			.replace(/\{conversation}/g, `<a target="_blank" class="external" href="${targetUrl}">${escapeHtml(displayName)} ↗</a>`),
-			{
-				isHTML: true,
-			})
+		{
+			isHTML: true,
+		})
 	} catch (exception) {
 		console.error('Error posting location to conversation', exception, exception.response?.status)
 		if (exception.response?.status === 403) {
@@ -57,34 +57,13 @@ function init() {
 	window.OCA.Maps.registerMapsAction({
 		label: t('spreed', 'Share to a conversation'),
 		icon: 'icon-talk',
-		callback: (location) => {
-			const container = document.createElement('div')
-			container.id = 'spreed-post-location-to-room-select'
-			const body = document.getElementById('body-user')
-			body.appendChild(container)
-
-			const RoomSelector = () => import('./components/RoomSelector.vue')
-			const vm = new Vue({
-				el: container,
-				render: h => h(RoomSelector, {
-					props: {
-						dialogTitle: t('spreed', 'Share to conversation'),
-						showPostableOnly: true,
-						isPlugin: true,
-					},
-				}),
+		callback: async (location) => {
+			const conversation = await requestRoomSelection('spreed-post-location-to-room-select', {
+				dialogTitle: t('spreed', 'Share to conversation'),
+				showPostableOnly: true,
 			})
 
-			vm.$root.$on('close', () => {
-				vm.$el.remove()
-				vm.$destroy()
-			})
-			vm.$root.$on('select', (conversation) => {
-				vm.$el.remove()
-				vm.$destroy()
-
-				postLocationToRoom(location, conversation)
-			})
+			postLocationToRoom(location, conversation)
 		},
 	})
 }
@@ -99,10 +78,5 @@ __webpack_nonce__ = btoa(getRequestToken())
 // We do not want the index.php since we're loading files
 // eslint-disable-next-line
 __webpack_public_path__ = generateFilePath('spreed', '', 'js/')
-
-Vue.prototype.t = t
-Vue.prototype.n = n
-Vue.prototype.OC = window.OC
-Vue.prototype.OCA = window.OCA
 
 document.addEventListener('DOMContentLoaded', init)
